Extract the home screen into its module
