+++ lib/main.dart
@@ -1,8 +1,8 @@
 import 'package:andax/config/themes.dart';
-import 'package:andax/screens/home_screen.dart';
+import 'package:andax/modules/home/screens/home.dart';
 import 'package:andax/store.dart';
 import 'package:flutter/material.dart';
-import 'widgets/splash_screen.dart';
+import 'modules/home/screens/splash.dart';
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
